labs/05-mcp/demo-03-mcp-weather-server/pyproject.toml
[project]
name = "demo-03-mcp-weather-server"
version = "0.1.0"
description = "MCP server integrating external Weather API to demonstrate real-world tool usage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
